+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Scripts.hs
@@ -4,6 +4,9 @@ module Cardano.Ledger.Api.Scripts (
   ScriptHash,
   CostModels (..),
   ValidityInterval (..),
+
+  -- * Tools
+  module Cardano.Ledger.Api.Scripts.Tools,
 )
 where
 
@@ -11,5 +14,6 @@ import Cardano.Ledger.Allegra.Scripts (ValidityInterval (..))
 import Cardano.Ledger.Alonzo.Scripts (CostModels (..))
 import Cardano.Ledger.Api.Era ()
 import Cardano.Ledger.Api.Scripts.Data
+import Cardano.Ledger.Api.Scripts.Tools
 import Cardano.Ledger.Core (EraScript (..))
 import Cardano.Ledger.Hashes (ScriptHash)